backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ollama==0.2.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
python-multipart==0.0.6